.NOW TIM: Wrong type for time.
A int was used, time is long long.

+++ MicroFrameworkPK_v4_3/DeviceCode/Targets/Native/STM32F10x/DeviceCode/drivers/Timer/advancedtimer/netmf_advancedtimer.cpp
@@ -230,7 +230,7 @@ DeviceStatus STM32F10x_AdvancedTimer::SetCompare(UINT64 compareValue)
 {
 	uint16_t tar_upper;
 	uint16_t now_upper;
-	uint32_t now;
+	uint64_t now;
 #if defined(DEBUG_EMOTE_ADVTIME)
 	volatile UINT64 NowTicks = g_STM32F10x_AdvancedTimer.Get64Counter();
 	if(NowTicks > compareValue) {
